#  A benchmark for plugin serialization and an optimization