Before:
```
-------------------------------------------------------------------------
Benchmark                                  Time           CPU Iterations
-------------------------------------------------------------------------
BM_PluginSerializationBenchmark   3366190962 ns 2683217200 ns          1   14.3157MB/s
```
After:
```
-----------------------------------------------------------------------
Benchmark                                Time           CPU Iterations
-----------------------------------------------------------------------
BM_PluginSerializationBenchmark 2367293571 ns 1638010500 ns          1   23.4505MB/s
```